Add indicators

As `f_remove_indicators()` but adds indicators back in. Obviously only indicators that were originally present in the input data can be added.

Usage

f_add_indicators(coin, add_indicators = NULL)

Arguments

coin

The coin

add_indicators

Character vector of indicator codes to add back in.

Value

coin Updated coin.

Examples

MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )
#> Removed indicators with no data points: A.A.1, A.M.5, A.M.6, A.V.5, S.T.3, C.C.2, C.C.3, C.C.4, C.J.1, C.J.3, C.S.1, C.S.2, C.S.3
#> Removed categories containing no indicators: Acc_Hum
#> iData checked and OK.
#> iMeta checked and OK.
#> Written data set to .$Data$Raw

# remove first
MVI <- f_remove_indicators(MVI, c("S.G.3", "A.M.1"))
#> coin has been regenerated using new specs.
# print method
MVI
#> --------------
#> A coin with...
#> --------------
#> Input:
#>   Units: 340 (GT0101, GT0102, GT0103, ...)
#>   Indicators: 39 (S.A.1, S.A.3, S.A.4, ...)
#>   Denominators: 0 ()
#>   Groups: 0 (none)
#> 
#> Structure:
#>   Level 1 Indicator: 39 indicators (A.D.1, A.M.2, A.M.4, ...) 
#>   Level 2 Category: 12 groups (Desastres, Desplaz, Violencia, ...) 
#>   Level 3 Dimension: 3 groups (Amenazas, Cap_Resp, Sit_SocEc) 
#>   Level 4 Index: 1 groups (MVI) 
#> 
#> Data sets:
#>   Raw (340 units)

# add one back
MVI <- f_add_indicators(MVI, "S.G.3")
#> coin has been regenerated using new specs.
MVI
#> --------------
#> A coin with...
#> --------------
#> Input:
#>   Units: 340 (GT0101, GT0102, GT0103, ...)
#>   Indicators: 40 (S.A.1, S.A.3, S.A.4, ...)
#>   Denominators: 0 ()
#>   Groups: 0 (none)
#> 
#> Structure:
#>   Level 1 Indicator: 40 indicators (A.D.1, A.M.2, A.M.4, ...) 
#>   Level 2 Category: 12 groups (Desastres, Desplaz, Violencia, ...) 
#>   Level 3 Dimension: 3 groups (Amenazas, Cap_Resp, Sit_SocEc) 
#>   Level 4 Index: 1 groups (MVI) 
#> 
#> Data sets:
#>   Raw (340 units)