Build index

This function builds the index. It assumes that at this point you have imported your data and built the coin. Also optionally you have analysed and possibly removed indicators, but taken no further steps.

f_build_index(
  coin,
  agg_method = "a_amean",
  max_winsorisation = 5,
  skew_thresh = 2,
  kurt_thresh = 3.5
)

Arguments

coin

The coin.

agg_method

One of `"a_amean"` (arithmetic mean), `"a_gmean"` (geometric mean), `"a_bod"` (benefit of doubt via Compind package) or `"a_wroclaw"` (Wroclaw Taxonomic Method via Compind).

max_winsorisation

default is 5,

skew_thresh

default is 2,

kurt_thresh

default is 3.5

Value

coin Updated coin.

Details

The operations performed here are:

- Treatment of outliers using a standard Winsorisation/log approach - Normalistion to the 1, 100 interval - Aggregation using one of the methods specified by `agg_method`

The output is an updated coin.

NOTE that for the Compind-powered aggregation methods, we need to be careful about how the aggregated values are rescaled or not in the case of a single indicator in an aggregation group. I initially blocked single-indicator groups here but relaxed this after adjusting the aggregation wrapper functions.

Examples

MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )
#> Removed indicators with no data points: A.A.1, A.M.5, A.M.6, A.V.5, S.T.3, C.C.2, C.C.3, C.C.4, C.J.1, C.J.3, C.S.1, C.S.2, C.S.3
#> Removed categories containing no indicators: Acc_Hum
#> iData checked and OK.
#> iMeta checked and OK.
#> Written data set to .$Data$Raw

MVI1 <- f_build_index(coin = MVI, 
                     agg_method = "a_amean", #   (arithmetic mean),
                     max_winsorisation = 5,
                     skew_thresh = 2,
                     kurt_thresh = 3.5)
#> Written data set to .$Data$Treated
#> Written data set to .$Data$Normalised
#> Written data set to .$Data$Aggregated

MVI2 <- f_build_index(coin = MVI, 
                     agg_method = "a_gmean", # (geometric mean),
                     max_winsorisation = 5,
                     skew_thresh = 2,
                     kurt_thresh = 3.5)
#> Written data set to .$Data$Treated
#> Written data set to .$Data$Normalised
#> Written data set to .$Data$Aggregated

MVI3 <- f_build_index(coin = MVI, 
                     agg_method = "a_bod", # (benefit of doubt via Compind package) 
                     max_winsorisation = 5,
                     skew_thresh = 2,
                     kurt_thresh = 3.5)
#> Written data set to .$Data$Treated
#> Written data set to .$Data$Normalised
#> Written data set to .$Data$Aggregated

MVI4 <- f_build_index(coin = MVI, 
                     agg_method = "a_wroclaw", # (Wroclaw Taxonomic Method via Compind)
                     max_winsorisation = 5,
                     skew_thresh = 2,
                     kurt_thresh = 3.5)
#> Written data set to .$Data$Treated
#> Written data set to .$Data$Normalised
#> Written data set to .$Data$Aggregated