Display results table

Outputs an interactive results table suitable for HTML documents and the app. .

f_display_results_table(coin, type = "scores")

Arguments

coin

COIN object, or list with first entry is the indicator metadata, second entry is the aggregation metadata

type

set type = "scores" or "ranks"

Value

a visualised data table

Examples

MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )
#> Removed indicators with no data points: A.A.1, A.M.5, A.M.6, A.V.5, S.T.3, C.C.2, C.C.3, C.C.4, C.J.1, C.J.3, C.S.1, C.S.2, C.S.3
#> Removed categories containing no indicators: Acc_Hum

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))

MVI2 <- f_build_index(coin = MVI)
#> Written data set to .$Data$Treated
#> Written data set to .$Data$Normalised
#> Written data set to .$Data$Aggregated
f_display_results_table(coin = MVI2)