Get last weights

Description

f_get_last_weights(coin)

Arguments

coin

COIN object, or list with first entry is the indicator metadata, second entry is the aggregation metadata

Value

w weight

Examples

MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )
#> Removed indicators with no data points: A.A.1, A.M.5, A.M.6, A.V.5, S.T.3, C.C.2, C.C.3, C.C.4, C.J.1, C.J.3, C.S.1, C.S.2, C.S.3
#> Removed categories containing no indicators: Acc_Hum

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))
MVI2 <- f_analyse_indicators(MVI)
MVI3 <- f_build_index(coin = MVI2)
#> Written data set to .$Data$Treated
#> Written data set to .$Data$Normalised
#> Written data set to .$Data$Aggregated
weight <- f_get_last_weights(MVI3)
weight 
#>          iCode Level Weight
#> 12       S.A.1     1      1
#> 13       S.A.3     1      1
#> 14       S.A.4     1      1
#> 30       C.C.1     1      1
#> 2        A.D.1     1      1
#> 3        A.M.1     1      1
#> 4        A.M.2     1      1
#> 5        A.M.4     1      1
#> 15       S.E.1     1      1
#> 16       S.E.2     1      1
#> 17       S.E.3     1      1
#> 34       C.E.4     1      1
#> 35       C.E.5     1      1
#> 36       C.E.6     1      1
#> 37       C.E.8     1      1
#> 38       C.E.9     1      1
#> 39      C.E.10     1      1
#> 40      C.E.11     1      1
#> 18       S.G.1     1      1
#> 19       S.G.2     1      1
#> 20       S.G.3     1      1
#> 21       S.G.4     1      1
#> 22       S.G.5     1      1
#> 23       S.G.6     1      1
#> 24       S.G.7     1      1
#> 25       S.G.8     1      1
#> 49       C.J.2     1      1
#> 54       C.S.4     1      1
#> 41       C.I.1     1      1
#> 42       C.I.2     1      1
#> 43       C.I.3     1      1
#> 44       C.I.4     1      1
#> 45       C.I.5     1      1
#> 46       C.I.6     1      1
#> 47       C.I.7     1      1
#> 26       S.T.1     1      1
#> 27       S.T.2     1      1
#> 29       S.T.4     1      1
#> 8        A.V.1     1      1
#> 9        A.V.2     1      1
#> 10       A.V.3     1      1
#> 56   Violencia     2      1
#> 58   Desastres     2      1
#> 63 Comunitaria     2      1
#> 64   Educacion     2      1
#> 65       Salud     2      1
#> 66    Serv_Pub     2      1
#> 67    Justicia     2      1
#> 55     Desplaz     2      1
#> 59     Trabajo     2      1
#> 60    Economia     2      1
#> 61   Alim_Vivi     2      1
#> 62    Grup_Vul     2      1
#> 68     Mov_Hum     3      1
#> 69    Amenazas     3      1
#> 70   Sit_SocEc     3      1
#> 71    Cap_Resp     3      1
#> 72         MVI     4      1