Plot map with main index and sub index

Plots an interactive choropleth map using supplied shape files and powered by the leaflet package. This second version of the summary map display the main severity index together with all its sub-dimensions each them presented as selected layers. The infobox present the index (or subindex) value together with the rank. The shapefile is automatically simplified so that the resulting html viz remains light.

Usage

f_plot_map2(
  coin,
  dset = "Aggregated",
  shp_path,
  shp_name,
  shp_ucode,
  dTolerance = 1000,
  palette = "inferno"
)

Arguments

coin

The coin

dset

Name of data set in the coin from which to extract the indicator/aggregate to plot

shp_path

Assume a shapefile matching with ucode Guatemala example is currently at `"./inst/shp/gtm_admbnda_adm2_ocha_conred_20190207.shp"`

shp_name

name of corresponding admin unit in the shapefile "ADM2_ES"

shp_ucode

name of corresponding admincode matching with ucode in the shapefile for instance `"ADM2_PCODE"` being the main key to use

dTolerance

1000 per default - parameter to simplify the shapefile so that the report is not too heavy...

palette

color palette for the map - selected as "inferno" per default to be inclusive of color blind people

Value

a leaflet object

Examples

MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )
#> Removed indicators with no data points: A.A.1, A.M.5, A.M.6, A.V.5, S.T.3, C.C.2, C.C.3, C.C.4, C.J.1, C.J.3, C.S.1, C.S.2, C.S.3
#> Removed categories containing no indicators: Acc_Hum
#> iData checked and OK.
#> iMeta checked and OK.
#> Written data set to .$Data$Raw

MVI <- f_build_index(MVI)
#> Written data set to .$Data$Treated
#> Written data set to .$Data$Normalised
#> Written data set to .$Data$Aggregated

shp_path <-  system.file("data_input/gtm_admbnda_adm2_ocha_conred_20190207.shp",
                                            package = "BuildIndex")

shp_name <- "ADM2_ES"
shp_ucode <- "ADM2_PCODE"

f_plot_map2(coin = MVI, 
           dset = "Aggregated",
           shp_path = shp_path,
           shp_name = shp_name,
           shp_ucode = shp_ucode )      
#> Loading required namespace: rgdal