Print Coin

This is a print-style text output function for summarising the contents of the coin.

f_print_coin(coin)

Arguments

coin

The coin

Value

A coin

Details

It is intended to be used immediately on loading data, to show the user what they have input, so they can check for any unexpected things.

Examples

MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )
#> Removed indicators with no data points: A.A.1, A.M.5, A.M.6, A.V.5, S.T.3, C.C.2, C.C.3, C.C.4, C.J.1, C.J.3, C.S.1, C.S.2, C.S.3
#> Removed categories containing no indicators: Acc_Hum
#> iData checked and OK.
#> iMeta checked and OK.
#> Written data set to .$Data$Raw

f_print_coin(MVI)
#> ----------
#> Your data:
#> ----------
#> Input:
#>   Units: 340 (GT0101, GT0102, GT0103, ...)
#>   Indicators: 41 (S.A.1, S.A.3, S.A.4, ...)
#> 
#> Structure:
#>   Level 1 Indicator: 41 indicators (A.D.1, A.M.1, A.M.2, ...) 
#>   Level 2 Category: 12 groups (Desastres, Desplaz, Violencia, ...) 
#>   Level 3 Dimension: 3 groups (Amenazas, Cap_Resp, Sit_SocEc) 
#>   Level 4 Index: 1 groups (MVI) 
#>