Interactive sunburst plot of index structure

Plots the structure of the index using a sunburst plot using **plotly**.

Usage

f_plot_framework(coin, seg_cols = NULL)

Arguments

seg_cols

A character vector of colour codes, one for each segment in the plot. The length of this vector must be equal to the number of segments, i.e. the sum of the number of indicators and aggregates in each level.

COIN

coin object

Value

fig as a plotly plot

Examples

MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )
#> Removed indicators with no data points: A.A.1, A.M.5, A.M.6, A.V.5, S.T.3, C.C.2, C.C.3, C.C.4, C.J.1, C.J.3, C.S.1, C.S.2, C.S.3
#> Removed categories containing no indicators: Acc_Hum
#> iData checked and OK.
#> iMeta checked and OK.
#> Written data set to .$Data$Raw

f_plot_framework(MVI)